pyproject.toml
[tool.poetry]
name = "profile-api"
version = "0.1.0"
description = ""
authors = ["pawlikslawomir@gmail.com"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
drf-spectacular = "^0.26.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"

[tool.poetry.group.dev.dependencies]
ipython = "^8.14.0"
flake8 = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

